Fix Serial Write
In order to guarantee the width of the start bit in a serial transfer we
need ot make sure that we are waiting for the correct amount of time.
Since the timer is already started we have no guarantee of the current
count value and therefore should add an initial wait.